requirements.txt
# SkyDash Terminal Admin Requirements

# Core terminal UI
rich>=13.0.0

# System monitoring (optional but recommended)
psutil>=5.9.0

# Configuration management
PyYAML>=6.0

# HTTP requests for API interactions
requests>=2.28.0

# PostgreSQL database adapter
psycopg2-binary>=2.9.0

# DNS resolution for email module
dnspython>=2.2.0

# Docker API client
docker>=6.0.0

# SSL/TLS certificate handling
cryptography>=3.4.0

# Additional useful libraries
python-dateutil>=2.8.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
